Add managesvc param to have the service managed (default) or not.
For networks with HA gateways/routers, it's not ideal to have many
radvd running concurrently in the same network segment. To support
those environments, make it possible for the radvd service be managed
by external tools.

Service's ensure => true/false are synonyms for running/stopped so
use $radvd::start directly.